feat(scopes): new scopeTargets option
This is a relatively small step as far as having a good
story around auth, but I think at least making it a bit
clearer what these options are and where they come from
is going to be nice.

At least opt-check doesn't have a ton of random @-opts now.