fix(tar): bump tar and bring back ignore-renaming
pacote@5 dropped support for a Very Legacy™ thing where npm would, on
extract, rename .gitignore to .npmignore if there was not already an
.npmignore.

This was primarily to support the use case where folks would use npm
install on github release tarballs.

Ref: https://github.com/npm/npm/issues/5658

BREAKING CHANGE: this reverts a previous change to disable this feature.